requirements.txt
litellm==1.66.3
google-adk==1.9.0